Create README.md
The project “Data Compression and Decompression” System is totally built on Java technology and provides an interface to the users so that they can easily store and transfer large files.
The layout of the project is built using the swing and AWT packages of Java using their predefined classes, interfaces, and methods. 
The io(Input Output) and util packages of Java provide the predefined classes to compress and decompress the files directly by passing the filename and location.
Using this project, the users can compress the data according to the requirement at any point and for any number of times.